로그인처리  - 로그인했을때 메뉴와 프로필이미지     정보와 직무에 대한 정보를 db에서 가져오기
