+++ WebContent/emp/login.jsp
@@ -45,30 +45,30 @@
 
 																		<tr>
 																			<td colspan=3><img
-																				src="/serverweb/images/page_login.gif"></td>
+																				src="/kimsaemERP/images/page_login.gif"></td>
 																		</tr>
 																		<tr>
 																			<td height=25></td>
 																		</tr>
 																		<tr>
-																			<td background="/serverweb/images/inputBg.gif"><img
-																				src="/serverweb/images/login_id.gif"></td>
-																			<td background="/serverweb/images/inputBg.gif"><input
+																			<td background="/kimsaemERP/images/inputBg.gif"><img
+																				src="/kimsaemERP/images/login_id.gif"></td>
+																			<td background="/kimsaemERP/images/inputBg.gif"><input
 																				type="text" name="id" size="15" class="grayinput"
 																				tabindex=1 style="width: 156;"
 																				 value="<%= id%>"></td>
 																			<td rowspan=3 style="padding-left: 10" valign=bottom>
 																			<input
-																				type='image' src="/serverweb/images/btn_login.gif"
+																				type='image' src="/kimsaemERP/images/btn_login.gif"
 																				border="0" tabindex=3></td>
 																		</tr>
 																		<tr>
 																			<td height=10></td>
 																		</tr>
 																		<tr>
-																			<td background="/serverweb/images/inputBg.gif"><img
-																				src="/serverweb/images/login_pass.gif"></td>
-																			<td background="/serverweb/images/inputBg.gif"><input
+																			<td background="/kimsaemERP/images/inputBg.gif"><img
+																				src="/kimsaemERP/images/login_pass.gif"></td>
+																			<td background="/kimsaemERP/images/inputBg.gif"><input
 																				type="password" name="pass" size="15"
 																				class="grayinput" tabindex='2' style='width: 156;'>
 																			</td>
