+++ src/fw/DBUtil.java
@@ -25,8 +25,8 @@ public class DBUtil {
 	public static Connection getConnect(){
 		Connection con = null;
 		String url = "jdbc:oracle:thin:@127.0.0.1:1521:xe";
-		String user = "scott";
-		String password = "tiger";
+		String user = "erp";
+		String password = "erp";
 		try{
 			con =DriverManager.getConnection(url, user, password);
 		}catch (SQLException e) {
